Create NLMPubmed_absscrape_months_sub100000abs
Core code to create biomedical document corpora (as part of broader NLM Pubmed abstract semantic space) based on scrape of all abstracts from each of the months of a given input year, up to a maximum of 100,000 scrapes per month in this version (the max allowed by the NLM eutils API). Corpora are saved locally in this version (with sample toy Windows path and directory) in two related forms: list of strings (each string i.e. document saved as a string) and nested list of tokens (each internal list a tokenization of a given abstract). The files are saved as simplejson text; the last two lines of code just show a sample of subsequent vectorization steps. NB: The use of the <abstract> XML tag as opposed to <abstracttext> as the delimiter for separating documents (i.e. individual abstracts) is essential to accurately scrape structured abstracts, which comprise a significant percentage of the NLM catalog. Structured abstracts split a full abstract into sections delimited by attribute markers, rendering the <abstracttext> tag unusable as a general delimiter. The <abstract> tag with additional string and parsed XML tag processing as above therefore serves this purpose and allows for generation of a "clean corpus" containing only abstract text, valuable for LSA and other document comparison approaches.